requirements.txt
flask==2.3.3
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
flask-login==0.6.3
flask-wtf==1.2.1
werkzeug==2.3.7
email-validator==2.1.0
flask-bcrypt==1.0.1
python-dotenv==1.0.0
wtforms==3.1.1
gunicorn==21.2.0
psycopg2-binary==2.9.9
fpdf2==2.7.6
qrcode==7.4.2
pillow==10.1.0